pyproject.toml
[tool.poetry]
name = "monkeypatch"
version = "0.1.0"
description = ""
authors = ["Ismael González Valverde <ismaelgv@ext.inditex.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "^1.55"
cryptography = "^3.4.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
